+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
 
-First stable release 01/11/2022 v1.0.0
+## [1.0.1] - 2022-11-08
+- Fixed: Session name changed to `filament_password_confirmed_at` to avoid conflicts with laravel RequirePassword middleware.  
+- Fixed: Password confirmation will not be requested on first regular login
+
+## [1.0.0] - 2022-11-01
+First stable release